# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "WonderBook API",
    "version": "1.0.0",
    "description": "Documentation complète WonderBook générée depuis les contrôleurs"
  },
  "servers": [
    { "url": "http://localhost:5000/api" }
  ],
  "paths": {
...
    "/books": {
      "get": {
        "tags": ["Books"],
        "summary": "Lister tous les livres",
        "responses": {
          "200": { "description": "Liste de livres renvoyée" }
        }
      },
      "post": {
        "tags": ["Books"],
        "summary": "Ajouter un livre",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "author": { "type": "string" },
                  "date": { "type": "string", "format": "date" },
                  "summary": { "type": "string" },
                  "cover_url": { "type": "string", "format": "uri" }
                },
                "required": ["title", "author"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Livre ajouté" },
          "400": { "description": "Erreur de validation" }
        }
      }
    },
...
    "/collection": {
      "get": {
        "tags": ["Collection"],
        "summary": "Obtenir la collection de l'utilisateur connecté",
        "responses": {
          "200": { "description": "Collection renvoyée" },
          "401": { "description": "Non autorisé" }
        }
      }
    },
...
    "/comments": {
      "post": {
        "tags": ["Comments"],
        "summary": "Ajouter un commentaire à un livre",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "bookId": { "type": "integer" },
                  "content": { "type": "string" },
                  "rating": { "type": "integer" }
                },
                "required": ["bookId", "content"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Commentaire ajouté" },
          "400": { "description": "Erreur de validation" }
        }
      }
    },
    "/topics": {
      "get": {
        "tags": ["Forum"],
        "summary": "Lister tous les topics",
        "responses": {
          "200": { "description": "Liste des topics renvoyée" }
        }
      },
      "post": {
        "tags": ["Forum"],
        "summary": "Créer un nouveau topic",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "content": { "type": "string" }
                },
                "required": ["title", "content"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Topic créé" }
        }
      }
    },
...
    "/posts": {
      "post": {
        "tags": ["Forum"],
        "summary": "Créer un post dans un topic",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "topicId": { "type": "integer" },
                  "content": { "type": "string" }
                },
                "required": ["topicId", "content"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Post ajouté" }
        }
      }
    },
...
  }
}